feature/go-1335 > Create "Other Income" page for admins/accountants

+++ rest/WEB-INF/src/org/bandahealth/idempiere/rest/service/db/InvoiceLineDBService.java
@@ -1,15 +1,17 @@
 package org.bandahealth.idempiere.rest.service.db;
 
 import org.adempiere.exceptions.AdempiereException;
+import org.bandahealth.idempiere.base.model.MBHBPSpecificPayerInfo;
+import org.bandahealth.idempiere.base.model.MBHPayerInfoFld;
 import org.bandahealth.idempiere.base.model.MCharge_BH;
+import org.bandahealth.idempiere.base.model.MOrderLine_BH;
 import org.bandahealth.idempiere.base.model.MProduct_BH;
+import org.bandahealth.idempiere.rest.model.BusinessPartnerSpecificPayerInformation;
 import org.bandahealth.idempiere.rest.model.Charge;
-import org.bandahealth.idempiere.rest.model.ExpenseCategory;
 import org.bandahealth.idempiere.rest.model.InvoiceLine;
+import org.bandahealth.idempiere.rest.model.OrderLine;
 import org.bandahealth.idempiere.rest.model.Product;
-import org.bandahealth.idempiere.rest.utils.DateUtil;
 import org.bandahealth.idempiere.rest.utils.StringUtil;
-import org.compiere.model.MElementValue;
 import org.compiere.model.MInvoiceLine;
 import org.compiere.model.Query;
 import org.compiere.util.Env;
@@ -17,6 +19,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -31,11 +35,15 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 	@Autowired
 	private ProductDBService productDBService;
 	@Autowired
-	private ExpenseCategoryDBService expenseCategoryDBService;
-	@Autowired
 	private AccountDBService accountDBService;
 	@Autowired
 	private ChargeDBService chargeDBService;
+	@Autowired
+	private OrderLineDBService orderLineDBService;
+	@Autowired
+	private BusinessPartnerSpecificPayerInformationDBService businessPartnerSpecificPayerInformationDBService;
+	@Autowired
+	private PayerInformationFieldDBService payerInformationFieldDBService;
 
 	@Override
 	public InvoiceLine saveEntity(InvoiceLine entity) {
@@ -49,17 +57,23 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 			invoiceLine.setC_Invoice_ID(entity.getInvoiceId());
 		}
 
+		if (entity.getOrderLineId() > 0) {
+			invoiceLine.setOrderLine(orderLineDBService.getEntityByIdFromDB(entity.getOrderLineId()));
+		} else if (entity.getOrderLine() != null && !StringUtil.isNullOrEmpty(entity.getOrderLine().getUuid())) {
+			invoiceLine.setOrderLine(orderLineDBService.getEntityByUuidFromDB(entity.getOrderLine().getUuid()));
+		}
+
 		if (entity.getCharge() != null && !StringUtil.isNullOrEmpty(entity.getCharge().getUuid())) {
 			MCharge_BH charge = chargeDBService.getEntityByUuidFromDB(entity.getCharge().getUuid());
 			if (charge != null) {
 				invoiceLine.setC_Charge_ID(charge.get_ID());
 			}
-		} else if (entity.getExpenseCategory() != null
-				&& !StringUtil.isNullOrEmpty(entity.getExpenseCategory().getUuid())) {
-			MCharge_BH charge = expenseCategoryDBService.getEntityByUuidFromDB(entity.getExpenseCategory().getUuid());
+		}
 
-			if (charge != null) {
-				invoiceLine.setC_Charge_ID(charge.get_ID());
+		if (entity.getProduct() != null) {
+			MProduct_BH product = productDBService.getEntityByUuidFromDB(entity.getProduct().getUuid());
+			if (product != null) {
+				invoiceLine.setM_Product_ID(product.get_ID());
 			}
 		}
 
@@ -67,14 +81,6 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 		// there
 		if (invoiceLine.getC_Charge_ID() == 0 && invoiceLine.getM_Product_ID() == 0) {
 			throw new AdempiereException("Invoice Line missing a charge or product");
-		}
-
-		if (entity.getProduct() != null) {
-			MProduct_BH product = productDBService.getEntityByUuidFromDB(entity.getProduct().getUuid());
-
-			if (product != null) {
-				invoiceLine.setM_Product_ID(product.get_ID());
-			}
 		}
 
 		if (entity.getPrice() != null) {
@@ -97,6 +103,36 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 		invoiceLine.setDescription(entity.getDescription());
 
 		invoiceLine.saveEx();
+		entity.setId(invoiceLine.get_ID());
+
+		// If there is any information to save with this line, save it
+		if (entity.getBusinessPartnerSpecificPayerInformationList() != null) {
+			entity.setBusinessPartnerSpecificPayerInformationList(
+					entity.getBusinessPartnerSpecificPayerInformationList().stream().map(
+							businessPartnerSpecificPayerInformation -> {
+								businessPartnerSpecificPayerInformation.setInvoiceLineId(entity.getId());
+								return businessPartnerSpecificPayerInformationDBService.saveEntity(
+										businessPartnerSpecificPayerInformation);
+							}).collect(Collectors.toList()));
+		} else {
+			entity.setBusinessPartnerSpecificPayerInformationList(new ArrayList<>());
+		}
+		// Delete what is no longer there
+		List<MBHBPSpecificPayerInfo> businessPartnerSpecificPayerInformationList =
+				businessPartnerSpecificPayerInformationDBService.getGroupsByIds(MBHBPSpecificPayerInfo::getC_InvoiceLine_ID,
+								MBHBPSpecificPayerInfo.COLUMNNAME_C_InvoiceLine_ID, Collections.singleton(entity.getId()))
+						.get(entity.getId());
+		if (businessPartnerSpecificPayerInformationList != null) {
+			businessPartnerSpecificPayerInformationList.stream()
+					.filter(
+							existingBusinessPartnerSpecificPayerInformation -> entity.getBusinessPartnerSpecificPayerInformationList()
+									.stream()
+									.noneMatch(
+											newBusinessPartnerSpecificPayerInformation -> newBusinessPartnerSpecificPayerInformation.getUuid()
+													.equals(existingBusinessPartnerSpecificPayerInformation.getBH_BP_Specific_Payer_Info_UU())))
+					.forEach(orderLineChargeInformation -> businessPartnerSpecificPayerInformationDBService
+							.deleteEntity(orderLineChargeInformation.getBH_BP_Specific_Payer_Info_UU()));
+		}
 
 		return createInstanceWithAllFields(getEntityByUuidFromDB(invoiceLine.getC_InvoiceLine_UU()));
 	}
@@ -110,28 +146,11 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 	protected InvoiceLine createInstanceWithAllFields(MInvoiceLine instance) {
 		try {
 			MProduct_BH product = productDBService.getEntityByIdFromDB(instance.getM_Product_ID());
+			InvoiceLine invoiceLine = new InvoiceLine(instance);
 			if (product != null) {
-				return new InvoiceLine(instance.getAD_Client_ID(), instance.getAD_Org_ID(),
-						instance.getC_InvoiceLine_UU(), instance.isActive(), DateUtil.parse(instance.getCreated()),
-						instance.getCreatedBy(), instance.getC_Invoice_ID(),
-						new Product(product.getName(), product.getM_Product_UU(), product), instance.getPriceActual(),
-						instance.getQtyInvoiced(), instance.getLineNetAmt(), instance.getDescription());
-			} else {
-				// check charge
-				MCharge_BH charge = expenseCategoryDBService.getEntityByIdFromDB(instance.getC_Charge_ID());
-				if (charge != null) {
-					MElementValue account = accountDBService.getEntityByIdFromDB(charge.getC_ElementValue_ID());
-					if (account != null) {
-						ExpenseCategory expenseCategory = new ExpenseCategory(charge.getC_Charge_UU(), charge.getName(),
-								charge.isBH_Locked(), account.getC_ElementValue_UU());
-						return new InvoiceLine(instance.getAD_Client_ID(), instance.getAD_Org_ID(),
-								instance.getC_InvoiceLine_UU(), instance.isActive(),
-								DateUtil.parse(instance.getCreated()), instance.getCreatedBy(), expenseCategory,
-								instance.getC_Invoice_ID(), instance.getPriceActual(), instance.getQtyInvoiced(),
-								instance.getLineNetAmt(), instance.getDescription());
-					}
-				}
+				invoiceLine.setProduct(new Product(product));
 			}
+			return invoiceLine;
 		} catch (Exception ex) {
 			log.severe(ex.getMessage());
 		}
@@ -146,7 +165,25 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 
 		// Batch call to get charges
 		Set<Integer> chargeIds = dbModels.stream().map(MInvoiceLine::getC_Charge_ID).collect(Collectors.toSet());
-		Map<Integer, MCharge_BH> charges = chargeDBService.getByIds(chargeIds);
+		Map<Integer, Charge> chargesById =
+				chargeDBService.transformData(new ArrayList<>(chargeDBService.getByIds(chargeIds).values())).stream()
+						.collect(Collectors.toMap(Charge::getId, charge -> charge));
+
+		// Batch call to get any order lines
+		Map<Integer, MOrderLine_BH> orderLinesById = orderLineDBService.getByIds(
+				dbModels.stream().map(MInvoiceLine::getC_OrderLine_ID).filter(orderLineId -> orderLineId > 0)
+						.collect(Collectors.toSet()));
+
+		// Batch call for insurance and donor payer information
+		Set<Integer> invoiceLineIds = dbModels.stream().map(MInvoiceLine::get_ID).collect(Collectors.toSet());
+		Map<Integer, List<MBHBPSpecificPayerInfo>> businessPartnerSpecificPayerInformationByInvoiceLineId =
+				businessPartnerSpecificPayerInformationDBService
+						.getGroupsByIds(MBHBPSpecificPayerInfo::getC_InvoiceLine_ID,
+								MBHBPSpecificPayerInfo.COLUMNNAME_C_InvoiceLine_ID, invoiceLineIds);
+		Map<Integer, MBHPayerInfoFld> payerInformationFieldsById = payerInformationFieldDBService.getByIds(
+				businessPartnerSpecificPayerInformationByInvoiceLineId.values().stream().flatMap(
+						businessPartnerSpecificPayerInformation -> businessPartnerSpecificPayerInformation.stream()
+								.map(MBHBPSpecificPayerInfo::getBH_Payer_Info_Fld_ID)).collect(Collectors.toSet()));
 
 		return dbModels.stream().map(invoiceLine -> {
 			InvoiceLine result = new InvoiceLine(invoiceLine);
@@ -154,15 +191,21 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 			if (products.containsKey(invoiceLine.getC_Invoice_ID())) {
 				result.setProduct(new Product(products.get(invoiceLine.getC_Invoice_ID())));
 			}
-
-			if (charges.containsKey(invoiceLine.getC_Charge_ID())) {
-				result.setCharge(new Charge(charges.get(invoiceLine.getC_Charge_ID())));
+			if (orderLinesById.containsKey(invoiceLine.getC_OrderLine_ID())) {
+				result.setOrderLine(new OrderLine(orderLinesById.get(invoiceLine.getC_OrderLine_ID())));
 			}
-			
-			// charge type
-			
-			
-			// account
+
+			if (chargesById.containsKey(invoiceLine.getC_Charge_ID())) {
+				result.setCharge(chargesById.get(invoiceLine.getC_Charge_ID()));
+			}
+			if (businessPartnerSpecificPayerInformationByInvoiceLineId.containsKey(result.getId())) {
+				result.setBusinessPartnerSpecificPayerInformationList(
+						businessPartnerSpecificPayerInformationByInvoiceLineId.get(result.getId()).stream()
+								.map(BusinessPartnerSpecificPayerInformation::new).peek(
+										businessPartnerSpecificPayerInformation -> businessPartnerSpecificPayerInformation.setPayerInformationFieldUuid(
+												payerInformationFieldsById.get(businessPartnerSpecificPayerInformation.getPayerInformationFieldId())
+														.getBH_Payer_Info_Fld_UU())).collect(Collectors.toList()));
+			}
 
 			return result;
 
@@ -177,7 +220,7 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 	public List<InvoiceLine> getInvoiceLinesByInvoiceId(int invoiceId) {
 		List<MInvoiceLine> invoiceLines = new Query(Env.getCtx(), MInvoiceLine.Table_Name,
 				MInvoiceLine.COLUMNNAME_C_Invoice_ID + "=?", null).setParameters(invoiceId).setOnlyActiveRecords(true)
-						.setClient_ID().list();
+				.setClient_ID().list();
 		return invoiceLines.stream().map(this::createInstanceWithDefaultFields).collect(Collectors.toList());
 	}
 
@@ -194,6 +237,20 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 
 		List<MInvoiceLine> invoiceLines = new Query(Env.getCtx(), MInvoiceLine.Table_Name, whereClause, null)
 				.setParameters(invoiceId).setClient_ID().list();
+
+		// Get the associated order line charge information and delete it
+		Set<Integer> invoiceLineIds = invoiceLines.stream().map(MInvoiceLine::getC_InvoiceLine_ID)
+				.collect(Collectors.toSet());
+		boolean wereChildrenDeletesSuccessful = businessPartnerSpecificPayerInformationDBService
+				.getGroupsByIds(MBHBPSpecificPayerInfo::getC_InvoiceLine_ID,
+						MBHBPSpecificPayerInfo.COLUMNNAME_C_InvoiceLine_ID, invoiceLineIds)
+				.values().stream().flatMap(Collection::stream)
+				.allMatch(businessPartnerChargeInformation -> businessPartnerSpecificPayerInformationDBService
+						.deleteEntity(businessPartnerChargeInformation.getBH_BP_Specific_Payer_Info_UU()));
+		if (!wereChildrenDeletesSuccessful) {
+			throw new AdempiereException("There was an error deleting information");
+		}
+
 		for (MInvoiceLine invoiceLine : invoiceLines) {
 			invoiceLine.deleteEx(false);
 		}
